keep only one new_context function that can be reused
